refactor proxy utils
