+++ gateway/src/app.service.ts
@@ -3,6 +3,6 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class AppService {
   getHello(): string {
-    return 'THis is AMRIT API GATEWAY!';
+    return 'This is AMRIT API GATEWAY!';
   }
 }
